gutenberg_backend/requirements.txt
Django==4.1.7
djangorestframework==3.14.0
requests==2.28.2
stop_words==2018.7.23
